+++ meta.go
@@ -1,3 +1,3 @@
 package ortfodb
 
-const Version = "1.2.0"
+const Version = "1.3.0"