+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - exporters init command: bootstrap a new exporter manifest file
+- exporters: [sprig](https://masterminds.github.io/sprig/) functions are now available in templates, along with shell-escape strings
 
 ### Fixed