+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - exporters init command: bootstrap a new exporter manifest file
 - exporters: [sprig](https://masterminds.github.io/sprig/) functions are now available in templates, along with shell-escape strings
 - env variables ORTFO_DEBUG and ORTFODB_DEBUG can now be used as alternatives to DEBUG to enable debug mode. as with DEBUG, the value must be "1" to enable debug mode
+- cloud exporter: config "name" to rename the uploaded file. defaults to the local file name
 
 ### Fixed