+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2024-04-16
+
 ### Added
 
 - exporters init command: bootstrap a new exporter manifest file
@@ -80,6 +82,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial release
 
 [Unreleased]: https://github.com/ortfo/db/compare/v1.2.0...HEAD
+[1.3.0]: https://github.com/ortfo/db/-/releases/tag/v1.3.0
 [1.2.0]: https://github.com/ortfo/db/compare/v1.1.0...v1.2.0
 [1.1.0]: https://github.com/ortfo/db/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/ortfo/db/compare/v0.3.2...v1.0.0
@@ -89,3 +92,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/ortfo/db/releases/tag/v0.2.0
 
 [//]: # (C3-2-DKAC:GGH:Rortfo/db:Tv{t})
+
+[unreleased]: https://github.com/ortfo/db/-/compare/v1.3.0...main